Add some more extensions to the defaults
These are all pretty benign extensions and we use them regularly anyway.
`ScopedTypeVariables` in particular is probably a must-have.
`GeneralizedNewtypeDeriving` should encourage more newtypes,
which are a "Good thing."
The others are mostly useful for interacting with libraries.
Since, most of the rest of the ecosystem uses MPTCs and generics.