Remove dependency that provided an orphan
We can roll our own logic that isn't based off of orphans,
and also get rid of a dependency in the process.

This shows one of the main problems with orphans.
These instances are defined in an unexposed module of a transitive dependency.
There's not even any documentation about the orphans existing.
We are forced to have them because they're orphans.

These were the instances in question:
https://github.com/frasertweedale/hs-jose/blob/55cc848d42388e90a721fa05a8c0f98a1eeebc43/src/Crypto/JOSE/Types/Orphans.hs#L44-L49